+++ benchmarks/versus-external-pools/fixed-poolifier.js
@@ -1,12 +1,12 @@
 // IMPORT LIBRARIES
 const { FixedThreadPool } = require('poolifier')
 // FINISH IMPORT LIBRARIES
-const size = Number(process.env.POOL_SIZE)
-const iterations = Number(process.env.NUM_ITERATIONS)
+const size = parseInt(process.env.POOL_SIZE)
+const iterations = parseInt(process.env.NUM_ITERATIONS)
 const data = {
   test: 'MYBENCH',
   taskType: process.env.TASK_TYPE,
-  taskSize: process.env.TASK_SIZE
+  taskSize: parseInt(process.env.TASK_SIZE)
 }
 
 const fixedPool = new FixedThreadPool(
@@ -20,7 +20,7 @@ async function run () {
     promises.push(fixedPool.execute(data))
   }
   await Promise.all(promises)
-  // eslint-disable-next-line no-process-exit
+  // eslint-disable-next-line n/no-process-exit
   process.exit()
 }