+++ benchmarks/versus-external-pools/fixed-threadwork.js
@@ -1,11 +1,11 @@
 // IMPORT LIBRARIES
 const threadPool = require('./pool-threadwork')
 // FINISH IMPORT LIBRARIES
-const iterations = Number(process.env.NUM_ITERATIONS)
+const iterations = parseInt(process.env.NUM_ITERATIONS)
 const data = {
   test: 'MYBENCH',
   taskType: process.env.TASK_TYPE,
-  taskSize: process.env.TASK_SIZE
+  taskSize: parseInt(process.env.TASK_SIZE)
 }
 
 async function run () {
@@ -14,7 +14,7 @@ async function run () {
     promises.push(threadPool.run(data))
   }
   await Promise.all(promises)
-  // eslint-disable-next-line no-process-exit
+  // eslint-disable-next-line n/no-process-exit
   process.exit()
 }