+++ benchmarks/versus-external-pools/fixed-workerpool.js
@@ -1,9 +1,13 @@
 // IMPORT LIBRARIES
 const workerpool = require('workerpool')
 // FINISH IMPORT LIBRARIES
-const size = Number(process.env.POOL_SIZE)
-const iterations = Number(process.env.NUM_ITERATIONS)
-const dataArray = ['MYBENCH', process.env.TASK_TYPE, process.env.TASK_SIZE]
+const size = parseInt(process.env.POOL_SIZE)
+const iterations = parseInt(process.env.NUM_ITERATIONS)
+const dataArray = [
+  'MYBENCH',
+  process.env.TASK_TYPE,
+  parseInt(process.env.TASK_SIZE)
+]
 
 const workerPool = workerpool.pool(
   './workers/workerpool/function-to-bench-worker.js',
@@ -20,7 +24,7 @@ async function run () {
     promises.push(workerPool.exec('functionToBench', dataArray))
   }
   await Promise.all(promises)
-  // eslint-disable-next-line no-process-exit
+  // eslint-disable-next-line n/no-process-exit
   process.exit()
 }