+++ benchmarks/versus-external-pools/pool-threadwork.js
@@ -4,6 +4,6 @@ const { ThreadPool } = require('threadwork')
 // IMPORT FUNCTION TO BENCH
 const functionToBench = require('./functions/function-to-bench')
 // FINISH IMPORT FUNCTION TO BENCH
-const size = Number(process.env.POOL_SIZE)
+const size = parseInt(process.env.POOL_SIZE)
 
 module.exports = new ThreadPool({ task: functionToBench, size })