+++ src/worker/abstract-worker.ts
@@ -88,7 +88,7 @@ export abstract class AbstractWorker<
       value.workerId !== undefined
     ) {
       // Here you will receive messages
-      if (this.opts.async) {
+      if (this.opts.async === true) {
         this.runInAsyncScope(this.runAsync.bind(this), this, fn, value)
       } else {
         this.runInAsyncScope(this.run.bind(this), this, fn, value)
@@ -143,7 +143,7 @@ export abstract class AbstractWorker<
   protected abstract sendToMainWorker (message: MessageValue<Response>): void
 
   /**
-   * Check to see if the worker should be terminated, because its living too long.
+   * Checks if the worker should be terminated, because its living too long.
    */
   protected checkAlive (): void {
     if (